LayoutTests/platform/ios-simulator/transitions/cubic-bezier-overflow-transform-expected.txt

CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box1'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box1'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box2'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box2'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box2'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box3'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box3'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box4'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box4'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box4'
PASS - "-webkit-transform.0" property for "box1" element at 0.5s saw something close to: -0.73135
PASS - "-webkit-transform.0" property for "box1" element at 0.95s saw something close to: -0.98481
FAIL - "-webkit-transform.0" property for "box2" element at 0.1s expected: 0.89101 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
FAIL - "-webkit-transform.0" property for "box2" element at 0.5s expected: 0.25882 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
PASS - "-webkit-transform.0" property for "box2" element at 0.95s saw something close to: -0.97437
FAIL - "-webkit-transform.0" property for "box3" element at 0.1s expected: 0.37461 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
PASS - "-webkit-transform.0" property for "box3" element at 0.5s saw something close to: -0.9205
FAIL - "-webkit-transform.0" property for "box4" element at 0.15s expected: -0.62932 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
FAIL - "-webkit-transform.0" property for "box4" element at 0.5s expected: 0.17365 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
FAIL - "-webkit-transform.0" property for "box4" element at 0.8s expected: 0.66913 but saw: matrix(-1, 0.00000000000000012246467991473532, -0.00000000000000012246467991473532, -1, 0, 0)
